refactor: refactor SessionInfo
